// packages/extension/src/types/provider.ts
export enum ProviderName {
  samara = "samara",
  radiant = "radiant",
}
// ...
export enum ProviderType {
  evm,
  radiant, // or rvm ?
}